:truck: Move SerialisableConfig to AzureSerialisableModel
