+++ data_safe_haven/utility/__init__.py
@@ -1,3 +1,4 @@
+from .azure_serialisable_model import AzureSerialisableModel
 from .directories import config_dir
 from .enums import DatabaseSystem, SoftwarePackageCategory
 from .file_reader import FileReader
@@ -7,6 +8,7 @@ from .types import PathType
 from .yaml_serialisable_model import YAMLSerialisableModel
 
 __all__ = [
+    "AzureSerialisableModel",
     "config_dir",
     "DatabaseSystem",
     "FileReader",
